refactor(api): split openai.py into models/attachments/deepseek/streaming modules
The original 1326-line openai.py is now split into focused submodules:
- models.py — pydantic models, model constants, _resolve_model, _finish_reason
- attachments.py — file attachment collection/validation/upload
- deepseek.py — retry logic, session management, HTTP helpers for DeepSeek
- streaming.py — SSE helpers, _collect_non_stream, _stream_openai, _stream_guard

openai.py retains lifespan, routes (health, list_models, chat_completions),
provider resolution, and re-exports all names so existing imports continue to work.

653 tests pass, ruff clean.
