+++ danyapi/api/attachments.py
@@ -1,0 +1,108 @@
+from __future__ import annotations
+
+import base64
+from dataclasses import dataclass
+
+from fastapi import HTTPException
+
+from ..deepseek.client import DeepSeekError
+from .deepseek import _deepseek_error_detail, _deepseek_status, _handle_account_error
+from .models import ChatCompletionRequest
+
+MAX_FILES_PER_REQUEST = 50
+MAX_FILE_SIZE = 100 * 1024 * 1024
+
+
+@dataclass
+class Attachment:
+    data: bytes
+    name: str
+    content_type: str
+    is_image: bool
+
+
+def _split_data_uri(uri: str) -> tuple[str, bytes]:
+    if not uri.startswith("data:"):
+        raise HTTPException(400, "image_url must be a data URI (data:<mime>;base64,...)")
+    meta, _, payload = uri[5:].partition(",")
+    content_type = meta.split(";", 1)[0] or "application/octet-stream"
+    try:
+        data = base64.b64decode(payload, validate=True)
+    except ValueError as exc:
+        raise HTTPException(400, "invalid base64 in image_url") from exc
+    return content_type, data
+
+
+def _collect_attachments(req: ChatCompletionRequest) -> list[Attachment]:
+    attachments: list[Attachment] = []
+    for msg in req.messages:
+        if not isinstance(msg.content, list):
+            continue
+        for item in msg.content:
+            if not isinstance(item, dict):
+                continue
+            if item.get("type") == "image_url":
+                image_url = item.get("image_url")
+                if isinstance(image_url, str):
+                    uri = image_url
+                elif isinstance(image_url, dict) and isinstance(image_url.get("url"), str):
+                    uri = image_url["url"]
+                else:
+                    raise HTTPException(400, "invalid image_url value")
+                content_type, data = _split_data_uri(uri)
+                name = f"image_{len(attachments)}.{content_type.split('/')[-1] or 'bin'}"
+                attachments.append(Attachment(data, name, content_type, True))
+    for f in req.files or []:
+        if not f.name or not f.content:
+            raise HTTPException(400, "each file needs name and base64 content")
+        try:
+            data = base64.b64decode(f.content)
+        except ValueError as exc:
+            raise HTTPException(400, f"invalid base64 in file {f.name}") from exc
+        attachments.append(Attachment(data, f.name, f.content_type or "application/octet-stream", f.content_type.startswith("image/")))
+    return attachments
+
+
+def _validate_attachments(attachments: list[Attachment], model_type: str) -> None:
+    if not attachments:
+        return
+    if len(attachments) > MAX_FILES_PER_REQUEST:
+        raise HTTPException(400, f"too many files: max {MAX_FILES_PER_REQUEST} per request")
+    for att in attachments:
+        if len(att.data) > MAX_FILE_SIZE:
+            raise HTTPException(400, f"file {att.name} exceeds 100 MB limit")
+    if model_type == "expert":
+        raise HTTPException(400, "deepseek-v4-pro does not support file attachments")
+    if model_type == "vision" and any(not att.is_image for att in attachments):
+        raise HTTPException(400, "deepseek-v4-vision accepts images only")
+
+
+async def _fresh_pow_upload_headers(account) -> dict:
+    try:
+        return await account.pow_upload.make_header(lambda: account.client.create_pow_challenge("/api/v0/file/upload_file"))
+    except DeepSeekError as exc:
+        _handle_account_error(account, exc)
+        raise HTTPException(_deepseek_status(exc), _deepseek_error_detail(exc)) from exc
+
+
+async def _upload_attachments(account, attachments: list[Attachment], model_type: str, thinking: bool) -> list[str]:
+    file_ids: list[str] = []
+    for att in attachments:
+        pow_headers = await _fresh_pow_upload_headers(account)
+        try:
+            info = await account.client.upload_file(
+                att.data,
+                att.name,
+                att.content_type,
+                model_type,
+                thinking_enabled=thinking,
+                pow_headers=pow_headers,
+            )
+        except DeepSeekError as exc:
+            _handle_account_error(account, exc)
+            raise HTTPException(_deepseek_status(exc), f"file upload failed: {exc}") from exc
+        file_id = info.get("id")
+        if not file_id:
+            raise HTTPException(502, f"file upload failed for {att.name}: no file id")
+        file_ids.append(file_id)
+    return file_ids
