+++ danyapi/api/deepseek.py
@@ -1,0 +1,274 @@
+from __future__ import annotations
+
+import asyncio
+import json
+import logging
+import random
+
+import httpx
+from fastapi import HTTPException
+
+from ..accounts import AccountPool, DeepSeekAccount
+from ..config import settings
+from ..deepseek.client import DeepSeekError, DeepSeekSession
+from ..deepseek.stream import IncrementalSSE, MessageReconstructor
+from .models import CONTEXT_LENGTH_STATUS, CONTINUE_PROMPT, INPUT_EXCEEDS_LIMIT
+
+log = logging.getLogger("danyapi.api")
+
+
+RETRYABLE_FINISH_REASONS = {
+    "expert_busy_use_default",
+    "parallel_chat_limit",
+    "server_busy",
+    "busy",
+}
+MAX_RETRIES = 5
+RETRY_BACKOFF_SEC = 1.0
+RETRY_BACKOFF_MAX_SEC = 8.0
+
+DEEPSEEK_AUTH_ERROR_CODES = {40001, 40002, 40003, 40012, 40029}
+
+
+async def _human_delay() -> None:
+    delay = random.uniform(settings.human_delay_min, settings.human_delay_max)
+    if delay > 0:
+        await asyncio.sleep(delay)
+
+
+def _is_retryable_hint(rec: MessageReconstructor) -> bool:
+    hint = rec.hint_error
+    return bool(hint and hint.get("finish_reason") in RETRYABLE_FINISH_REASONS)
+
+
+RETRYABLE_HTTP_STATUSES = {408, 425, 429, 500, 502, 503, 504}
+
+
+def _is_retryable_http(exc: HTTPException) -> bool:
+    return exc.status_code in RETRYABLE_HTTP_STATUSES
+
+
+def _is_context_limit(rec: MessageReconstructor) -> bool:
+    if rec.status == CONTEXT_LENGTH_STATUS:
+        return True
+    hint = rec.hint_error
+    return bool(hint and hint.get("finish_reason") == CONTEXT_LENGTH_STATUS)
+
+
+def _is_input_exceeds_limit(rec: MessageReconstructor) -> bool:
+    if rec.status == INPUT_EXCEEDS_LIMIT:
+        return True
+    hint = rec.hint_error
+    return bool(hint and hint.get("finish_reason") == INPUT_EXCEEDS_LIMIT)
+
+
+def _input_exceeds_hint_from_http(exc: HTTPException) -> dict | None:
+    detail = exc.detail
+    if isinstance(detail, str):
+        try:
+            detail = json.loads(detail)
+        except json.JSONDecodeError:
+            return None
+    if not isinstance(detail, dict):
+        return None
+    if detail.get("finish_reason") != INPUT_EXCEEDS_LIMIT:
+        return None
+    message = detail.get("message")
+    return {
+        "message": message if isinstance(message, str) else "Content is too long",
+        "finish_reason": INPUT_EXCEEDS_LIMIT,
+    }
+
+
+def _drop_session(pool, account, session_key) -> None:
+    pool.forget(session_key)
+    pool.forget_context(session_key)
+    account.sessions.forget(session_key)
+
+
+def _deepseek_status(exc: DeepSeekError) -> int:
+    return 401 if exc.biz_code in DEEPSEEK_AUTH_ERROR_CODES else 502
+
+
+async def _send_with_auth(account, *args, **kwargs):
+    try:
+        return await _send_completion(*args, **kwargs)
+    except HTTPException as exc:
+        if exc.status_code in (401, 403):
+            account.mark_broken()
+        raise
+
+
+def _deepseek_error_detail(exc: DeepSeekError) -> str:
+    if exc.biz_code in DEEPSEEK_AUTH_ERROR_CODES:
+        return f"DeepSeek auth error: {exc}"
+    return f"DeepSeek error: {exc}"
+
+
+def _handle_account_error(account: DeepSeekAccount, exc: Exception) -> None:
+    code = getattr(exc, "biz_code", None)
+    if code in DEEPSEEK_AUTH_ERROR_CODES:
+        account.mark_broken()
+        log.warning("account #%d auth error %s: %s", account.index, code, exc)
+    else:
+        log.warning("account #%d error: %s", account.index, exc)
+
+
+async def _fresh_pow_headers(account) -> dict:
+    try:
+        return await account.pow.make_header(account.client.create_pow_challenge)
+    except DeepSeekError as exc:
+        _handle_account_error(account, exc)
+        raise HTTPException(_deepseek_status(exc), _deepseek_error_detail(exc)) from exc
+
+
+def _busy_error_body(rec: MessageReconstructor) -> str:
+    hint = rec.hint_error or {}
+    return json.dumps(
+        {
+            "error": {
+                "message": hint.get("message") or "DeepSeek server is busy, try again later",
+                "finish_reason": hint.get("finish_reason"),
+            }
+        },
+        ensure_ascii=False,
+    )
+
+
+async def _try_stop_stream(client, session_id: str, message_id: str | None) -> None:
+    if not session_id or not message_id:
+        return
+    try:
+        await client.stop_stream(session_id, message_id)
+    except Exception as exc:
+        log.debug("stop_stream failed for %s: %s", session_id, exc)
+
+
+async def _collect_continuation(
+    account,
+    session,
+    parent_message_id,
+    model_type,
+    thinking,
+    search,
+    ref_file_ids=None,
+) -> MessageReconstructor | None:
+    for attempt in range(MAX_RETRIES + 1):
+        if attempt:
+            await asyncio.sleep(min(RETRY_BACKOFF_SEC * (2 ** (attempt - 1)), RETRY_BACKOFF_MAX_SEC))
+        try:
+            pow_headers = await _fresh_pow_headers(account)
+            resp = await _send_with_auth(
+                account,
+                account.client,
+                pow_headers,
+                session.id,
+                parent_message_id,
+                CONTINUE_PROMPT,
+                model_type,
+                thinking,
+                search,
+                ref_file_ids,
+            )
+        except HTTPException as exc:
+            if _is_retryable_http(exc) and attempt < MAX_RETRIES:
+                log.warning(
+                    "deepseek continuation error (%s), attempt %d/%d",
+                    exc.status_code,
+                    attempt + 1,
+                    MAX_RETRIES,
+                )
+                continue
+            return None
+        rec = MessageReconstructor()
+        incremental = IncrementalSSE()
+        try:
+            async for chunk in resp.aiter_bytes():
+                for event in incremental.feed(chunk):
+                    rec.handle(event)
+            for event in incremental.finish():
+                rec.handle(event)
+        finally:
+            await resp.aclose()
+        if not (rec.content or rec.reasoning) and _is_retryable_hint(rec) and attempt < MAX_RETRIES:
+            continue
+        return rec
+
+
+async def _prepare_session(
+    account: DeepSeekAccount,
+    pool: AccountPool,
+    existing_sid: str | None,
+    context_seq: tuple[str, ...] | None = None,
+) -> tuple[DeepSeekSession, str, str | None]:
+    try:
+        session, session_key = await account.sessions.obtain(existing_sid)
+    except DeepSeekError as exc:
+        _handle_account_error(account, exc)
+        raise HTTPException(_deepseek_status(exc), _deepseek_error_detail(exc)) from exc
+    if session_key != existing_sid:
+        pool.register(account.index, session_key)
+        if existing_sid:
+            pool.forget(existing_sid)
+            pool.forget_context(existing_sid)
+            account.sessions.forget(existing_sid)
+    if context_seq:
+        pool.index_context(session_key, context_seq)
+    return session, session_key, session.last_message_id
+
+
+async def _send_completion(
+    client,
+    pow_headers,
+    session_id,
+    parent_message_id,
+    prompt,
+    model_type,
+    thinking,
+    search,
+    ref_file_ids=None,
+):
+    try:
+        resp = await client.completion(
+            chat_session_id=session_id,
+            prompt=prompt,
+            parent_message_id=parent_message_id,
+            model_type=model_type,
+            thinking_enabled=thinking,
+            search_enabled=search,
+            ref_file_ids=ref_file_ids,
+            pow_headers=pow_headers,
+        )
+    except httpx.HTTPStatusError as exc:
+        raise HTTPException(exc.response.status_code, exc.response.text[:500]) from exc
+    except httpx.HTTPError as exc:
+        raise HTTPException(502, f"DeepSeek request failed: {exc}") from exc
+
+    if resp.status_code != 200:
+        body = await resp.aread()
+        await resp.aclose()
+        raise HTTPException(resp.status_code, body[:500].decode("utf-8", errors="replace"))
+
+    content_type = resp.headers.get("content-type", "")
+    if "text/event-stream" not in content_type:
+        body = await resp.aread()
+        await resp.aclose()
+        try:
+            payload = json.loads(body)
+        except json.JSONDecodeError as exc:
+            raise HTTPException(502, body[:500].decode("utf-8", errors="replace")) from exc
+        data = payload.get("data") or {}
+        if data.get("biz_code"):
+            code = data["biz_code"]
+            status = 401 if code in DEEPSEEK_AUTH_ERROR_CODES else 502
+            raise HTTPException(status, f"DeepSeek error {code}: {data.get('biz_msg')}")
+        if payload.get("code"):
+            code = payload["code"]
+            status = 401 if code in DEEPSEEK_AUTH_ERROR_CODES else 502
+            raise HTTPException(
+                status,
+                f"DeepSeek error {code}: {payload.get('msg') or payload.get('message')}",
+            )
+        raise HTTPException(502, "unexpected non-stream response")
+    return resp
+
