+++ danyapi/api/openai.py
@@ -1,106 +1,26 @@
 from __future__ import annotations
 
-import asyncio
-import base64
-import json
 import logging
-import random
-import sys
-import time
-import uuid
 from contextlib import asynccontextmanager
-from dataclasses import dataclass
 from typing import Any
 
-import httpx
 from fastapi import FastAPI, HTTPException
 from fastapi.responses import StreamingResponse
-from pydantic import BaseModel, Field
 
 from .. import tools as toolemu
-from ..accounts import AccountPool, AccountPoolBusy, DeepSeekAccount, account_lock
+from ..accounts import AccountPool, AccountPoolBusy, DeepSeekAccount
 from ..config import settings
-from ..deepseek.client import DeepSeekClient, DeepSeekError, DeepSeekSession
-from ..deepseek.stream import IncrementalSSE, MessageReconstructor
+from ..deepseek.client import DeepSeekClient
 from ..qwen import api as qwen_api
 from ..qwen.accounts import QwenAccount
 from ..qwen.client import QwenClient, QwenError
 from ..store import JsonStore
 
+# Import ChatCompletionRequest early so it's available for route signatures.
+# The full re-export block at the bottom brings in every other name used by tests.
+from .models import ChatCompletionRequest
+
 log = logging.getLogger("danyapi.api")
-
-MODEL_TYPE_BY_NAME = {
-    "deepseek-v4-flash": "default",
-    "deepseek-v4-pro": "expert",
-    "deepseek-v4-vision": "vision",
-}
-
-QWEN_DEFAULT_MODELS = [
-    {
-        "id": "qwen3.8-max",
-        "name": "Qwen3.8-Max",
-        "owned_by": "qwen",
-        "model_type": "chat",
-    },
-    {
-        "id": "qwen3.7-plus",
-        "name": "Qwen3.7-Plus",
-        "owned_by": "qwen",
-        "model_type": "chat",
-    },
-    {
-        "id": "qwen3.7-max",
-        "name": "Qwen3.7-Max",
-        "owned_by": "qwen",
-        "model_type": "chat",
-    },
-]
-
-STATUS_TO_FINISH_REASON = {
-    "FINISHED": "stop",
-    "CONTEXT_LENGTH_EXCEEDED": "length",
-    "CONTENT_FILTER": "content_filter",
-    "INCOMPLETE": "stop",
-    "WIP": "stop",
-    "TIMEOUT": "stop",
-}
-
-CONTEXT_LENGTH_STATUS = "CONTEXT_LENGTH_EXCEEDED"
-INPUT_EXCEEDS_LIMIT = "input_exceeds_limit"
-CONTINUE_PROMPT = "Continue"
-MAX_CONTINUE_ROUNDS = 5
-
-
-class ChatMessage(BaseModel):
-    role: str = "user"
-    content: Any = ""
-    tool_calls: list[Any] | None = None
-    tool_call_id: str | None = None
-    name: str | None = None
-
-
-class FileSpec(BaseModel):
-    name: str
-    content: str
-    content_type: str = "application/octet-stream"
-
-
-class ChatCompletionRequest(BaseModel):
-    model: str = Field(default="deepseek-v4-flash")
-    messages: list[ChatMessage] = Field(default_factory=list)
-    stream: bool = False
-    temperature: float | None = None
-    top_p: float | None = None
-    thinking: bool | None = None
-    search: bool | None = None
-    session_id: str | None = None
-    user: str | None = None
-    files: list[FileSpec] | None = None
-    tools: list[Any] | None = None
-    tool_choice: Any = None
-    parallel_tool_calls: bool | None = None
-    response_format: Any = None
-    stream_options: Any = None
 
 
 @asynccontextmanager
@@ -230,118 +150,6 @@ async def _fetch_qwen_models(client: QwenClient) -> list[dict]:
 app = FastAPI(title="DanyAPI", lifespan=lifespan)
 
 
-MAX_FILES_PER_REQUEST = 50
-MAX_FILE_SIZE = 100 * 1024 * 1024
-
-
-@dataclass
-class Attachment:
-    data: bytes
-    name: str
-    content_type: str
-    is_image: bool
-
-
-def _split_data_uri(uri: str) -> tuple[str, bytes]:
-    if not uri.startswith("data:"):
-        raise HTTPException(400, "image_url must be a data URI (data:<mime>;base64,...)")
-    meta, _, payload = uri[5:].partition(",")
-    content_type = meta.split(";", 1)[0] or "application/octet-stream"
-    try:
-        data = base64.b64decode(payload, validate=True)
-    except ValueError as exc:
-        raise HTTPException(400, "invalid base64 in image_url") from exc
-    return content_type, data
-
-
-def _collect_attachments(req: ChatCompletionRequest) -> list[Attachment]:
-    attachments: list[Attachment] = []
-    for msg in req.messages:
-        if not isinstance(msg.content, list):
-            continue
-        for item in msg.content:
-            if not isinstance(item, dict):
-                continue
-            if item.get("type") == "image_url":
-                image_url = item.get("image_url")
-                if isinstance(image_url, str):
-                    uri = image_url
-                elif isinstance(image_url, dict) and isinstance(image_url.get("url"), str):
-                    uri = image_url["url"]
-                else:
-                    raise HTTPException(400, "invalid image_url value")
-                content_type, data = _split_data_uri(uri)
-                name = f"image_{len(attachments)}.{content_type.split('/')[-1] or 'bin'}"
-                attachments.append(Attachment(data, name, content_type, True))
-    for f in req.files or []:
-        if not f.name or not f.content:
-            raise HTTPException(400, "each file needs name and base64 content")
-        try:
-            data = base64.b64decode(f.content)
-        except ValueError as exc:
-            raise HTTPException(400, f"invalid base64 in file {f.name}") from exc
-        attachments.append(Attachment(data, f.name, f.content_type or "application/octet-stream", f.content_type.startswith("image/")))
-    return attachments
-
-
-def _validate_attachments(attachments: list[Attachment], model_type: str) -> None:
-    if not attachments:
-        return
-    if len(attachments) > MAX_FILES_PER_REQUEST:
-        raise HTTPException(400, f"too many files: max {MAX_FILES_PER_REQUEST} per request")
-    for att in attachments:
-        if len(att.data) > MAX_FILE_SIZE:
-            raise HTTPException(400, f"file {att.name} exceeds 100 MB limit")
-    if model_type == "expert":
-        raise HTTPException(400, "deepseek-v4-pro does not support file attachments")
-    if model_type == "vision" and any(not att.is_image for att in attachments):
-        raise HTTPException(400, "deepseek-v4-vision accepts images only")
-
-
-async def _fresh_pow_upload_headers(account) -> dict:
-    try:
-        return await account.pow_upload.make_header(lambda: account.client.create_pow_challenge("/api/v0/file/upload_file"))
-    except DeepSeekError as exc:
-        _handle_account_error(account, exc)
-        raise HTTPException(_deepseek_status(exc), _deepseek_error_detail(exc)) from exc
-
-
-async def _upload_attachments(account, attachments: list[Attachment], model_type: str, thinking: bool) -> list[str]:
-    file_ids: list[str] = []
-    for att in attachments:
-        pow_headers = await _fresh_pow_upload_headers(account)
-        try:
-            info = await account.client.upload_file(
-                att.data,
-                att.name,
-                att.content_type,
-                model_type,
-                thinking_enabled=thinking,
-                pow_headers=pow_headers,
-            )
-        except DeepSeekError as exc:
-            _handle_account_error(account, exc)
-            raise HTTPException(_deepseek_status(exc), f"file upload failed: {exc}") from exc
-        file_id = info.get("id")
-        if not file_id:
-            raise HTTPException(502, f"file upload failed for {att.name}: no file id")
-        file_ids.append(file_id)
-    return file_ids
-
-
-def _resolve_model(model: str) -> str:
-    model_type = MODEL_TYPE_BY_NAME.get(model)
-    if model_type is None:
-        raise HTTPException(404, f"Unknown model: {model}")
-    return model_type
-
-
-def _finish_reason(status: Any) -> str:
-    if isinstance(status, str):
-        return STATUS_TO_FINISH_REASON.get(status, "stop")
-    return "stop"
-
-
 def _pool_stats(pool) -> dict | None:
     if pool is None:
         return None
@@ -391,25 +199,6 @@ async def list_models() -> dict:
     return {"object": "list", "data": models}
 
 
-RETRYABLE_FINISH_REASONS = {
-    "expert_busy_use_default",
-    "parallel_chat_limit",
-    "server_busy",
-    "busy",
-}
-MAX_RETRIES = 5
-RETRY_BACKOFF_SEC = 1.0
-RETRY_BACKOFF_MAX_SEC = 8.0
-
-DEEPSEEK_AUTH_ERROR_CODES = {40001, 40002, 40003, 40012, 40029}
-
-
-async def _human_delay() -> None:
-    delay = random.uniform(settings.human_delay_min, settings.human_delay_max)
-    if delay > 0:
-        await asyncio.sleep(delay)
-
-
 def _resolve_provider(model: str) -> str:
     if model.startswith("qwen"):
         return "qwen"
@@ -433,48 +222,6 @@ async def _acquire_account(pool: AccountPool, session_id: str | None):
         raise HTTPException(429, "all accounts are busy, try again later") from None
     except RuntimeError as exc:
         raise HTTPException(503, str(exc)) from exc
-
-
-def _include_usage(req: ChatCompletionRequest) -> bool:
-    opts = getattr(req, "stream_options", None)
-    return bool((opts or {}).get("include_usage"))
-
-
-def _deepseek_usage(total: int) -> dict:
-    value = max(0, int(total or 0))
-    return {"prompt_tokens": 0, "completion_tokens": value, "total_tokens": value}
-
-
-def _sse(data: dict) -> str:
-    return f"data: {json.dumps(data, ensure_ascii=False)}\n\n"
-
-
-async def _stream_guard(gen, model: str):
-    try:
-        async for item in gen:
-            yield item
-    except AccountPoolBusy:
-        chunk_id = f"chatcmpl-{uuid.uuid4().hex}"
-        created = int(time.time())
-        yield _sse(
-            {
-                "id": chunk_id,
-                "object": "chat.completion.chunk",
-                "created": created,
-                "model": model,
-                "error": {"message": "all accounts are busy, try again later"},
-                "choices": [{"index": 0, "delta": {}, "finish_reason": "error"}],
-            }
-        )
-        yield _sse(
-            {
-                "id": chunk_id,
-                "session_id": None,
-                "object": "chat.completion.chunk",
-                "choices": [],
-            }
-        )
-        yield "data: [DONE]\n\n"
 
 
 async def _chat_completions_deepseek(req: ChatCompletionRequest) -> Any:
@@ -591,736 +338,59 @@ async def _chat_completions_qwen(req: ChatCompletionRequest) -> Any:
         raise HTTPException(429, "all accounts are busy, try again later") from None
 
 
-async def _prepare_session(
-    account: DeepSeekAccount,
-    pool: AccountPool,
-    existing_sid: str | None,
-    context_seq: tuple[str, ...] | None = None,
-) -> tuple[DeepSeekSession, str, str | None]:
-    try:
-        session, session_key = await account.sessions.obtain(existing_sid)
-    except DeepSeekError as exc:
-        _handle_account_error(account, exc)
-        raise HTTPException(_deepseek_status(exc), _deepseek_error_detail(exc)) from exc
-    if session_key != existing_sid:
-        pool.register(account.index, session_key)
-        if existing_sid:
-            pool.forget(existing_sid)
-            pool.forget_context(existing_sid)
-            account.sessions.forget(existing_sid)
-    if context_seq:
-        pool.index_context(session_key, context_seq)
-    return session, session_key, session.last_message_id
+# ---------------------------------------------------------------------------
+# Re-exports: submodules of this package (models / attachments / deepseek /
+# streaming). Kept here so `danyapi.api.openai` exposes the same names as
+# before the split, and module-level monkeypatching in tests keeps working.
+# ---------------------------------------------------------------------------
 
-
-async def _send_completion(
-    client,
-    pow_headers,
-    session_id,
-    parent_message_id,
-    prompt,
-    model_type,
-    thinking,
-    search,
-    ref_file_ids=None,
-):
-    try:
-        resp = await client.completion(
-            chat_session_id=session_id,
-            prompt=prompt,
-            parent_message_id=parent_message_id,
-            model_type=model_type,
-            thinking_enabled=thinking,
-            search_enabled=search,
-            ref_file_ids=ref_file_ids,
-            pow_headers=pow_headers,
-        )
-    except httpx.HTTPStatusError as exc:
-        raise HTTPException(exc.response.status_code, exc.response.text[:500]) from exc
-    except httpx.HTTPError as exc:
-        raise HTTPException(502, f"DeepSeek request failed: {exc}") from exc
-
-    if resp.status_code != 200:
-        body = await resp.aread()
-        await resp.aclose()
-        raise HTTPException(resp.status_code, body[:500].decode("utf-8", errors="replace"))
-
-    content_type = resp.headers.get("content-type", "")
-    if "text/event-stream" not in content_type:
-        body = await resp.aread()
-        await resp.aclose()
-        try:
-            payload = json.loads(body)
-        except json.JSONDecodeError as exc:
-            raise HTTPException(502, body[:500].decode("utf-8", errors="replace")) from exc
-        data = payload.get("data") or {}
-        if data.get("biz_code"):
-            code = data["biz_code"]
-            status = 401 if code in DEEPSEEK_AUTH_ERROR_CODES else 502
-            raise HTTPException(status, f"DeepSeek error {code}: {data.get('biz_msg')}")
-        if payload.get("code"):
-            code = payload["code"]
-            status = 401 if code in DEEPSEEK_AUTH_ERROR_CODES else 502
-            raise HTTPException(
-                status,
-                f"DeepSeek error {code}: {payload.get('msg') or payload.get('message')}",
-            )
-        raise HTTPException(502, "unexpected non-stream response")
-    return resp
-
-
-def _is_retryable_hint(rec: MessageReconstructor) -> bool:
-    hint = rec.hint_error
-    return bool(hint and hint.get("finish_reason") in RETRYABLE_FINISH_REASONS)
-
-
-RETRYABLE_HTTP_STATUSES = {408, 425, 429, 500, 502, 503, 504}
-
-
-def _is_retryable_http(exc: HTTPException) -> bool:
-    return exc.status_code in RETRYABLE_HTTP_STATUSES
-
-
-def _is_context_limit(rec: MessageReconstructor) -> bool:
-    if rec.status == CONTEXT_LENGTH_STATUS:
-        return True
-    hint = rec.hint_error
-    return bool(hint and hint.get("finish_reason") == CONTEXT_LENGTH_STATUS)
-
-
-def _is_input_exceeds_limit(rec: MessageReconstructor) -> bool:
-    if rec.status == INPUT_EXCEEDS_LIMIT:
-        return True
-    hint = rec.hint_error
-    return bool(hint and hint.get("finish_reason") == INPUT_EXCEEDS_LIMIT)
-
-
-def _input_exceeds_hint_from_http(exc: HTTPException) -> dict | None:
-    detail = exc.detail
-    if isinstance(detail, str):
-        try:
-            detail = json.loads(detail)
-        except json.JSONDecodeError:
-            return None
-    if not isinstance(detail, dict):
-        return None
-    if detail.get("finish_reason") != INPUT_EXCEEDS_LIMIT:
-        return None
-    message = detail.get("message")
-    return {
-        "message": message if isinstance(message, str) else "Content is too long",
-        "finish_reason": INPUT_EXCEEDS_LIMIT,
-    }
-
-
-def _drop_session(pool, account, session_key) -> None:
-    pool.forget(session_key)
-    pool.forget_context(session_key)
-    account.sessions.forget(session_key)
-
-
-def _deepseek_status(exc: DeepSeekError) -> int:
-    return 401 if exc.biz_code in DEEPSEEK_AUTH_ERROR_CODES else 502
-
-
-async def _send_with_auth(account, *args, **kwargs):
-    try:
-        return await _send_completion(*args, **kwargs)
-    except HTTPException as exc:
-        if exc.status_code in (401, 403):
-            account.mark_broken()
-        raise
-
-
-def _deepseek_error_detail(exc: DeepSeekError) -> str:
-    if exc.biz_code in DEEPSEEK_AUTH_ERROR_CODES:
-        return f"DeepSeek auth error: {exc}"
-    return f"DeepSeek error: {exc}"
-
-
-def _handle_account_error(account: DeepSeekAccount, exc: Exception) -> None:
-    code = getattr(exc, "biz_code", None)
-    if code in DEEPSEEK_AUTH_ERROR_CODES:
-        account.mark_broken()
-        log.warning("account #%d auth error %s: %s", account.index, code, exc)
-    else:
-        log.warning("account #%d error: %s", account.index, exc)
-
-
-async def _fresh_pow_headers(account) -> dict:
-    try:
-        return await account.pow.make_header(account.client.create_pow_challenge)
-    except DeepSeekError as exc:
-        _handle_account_error(account, exc)
-        raise HTTPException(_deepseek_status(exc), _deepseek_error_detail(exc)) from exc
-
-
-def _busy_error_body(rec: MessageReconstructor) -> str:
-    hint = rec.hint_error or {}
-    return json.dumps(
-        {
-            "error": {
-                "message": hint.get("message") or "DeepSeek server is busy, try again later",
-                "finish_reason": hint.get("finish_reason"),
-            }
-        },
-        ensure_ascii=False,
-    )
-
-
-async def _try_stop_stream(client, session_id: str, message_id: str | None) -> None:
-    if not session_id or not message_id:
-        return
-    try:
-        await client.stop_stream(session_id, message_id)
-    except Exception as exc:
-        log.debug("stop_stream failed for %s: %s", session_id, exc)
-
-
-async def _collect_continuation(
-    account,
-    session,
-    parent_message_id,
-    model_type,
-    thinking,
-    search,
-    ref_file_ids=None,
-) -> MessageReconstructor | None:
-    for attempt in range(MAX_RETRIES + 1):
-        if attempt:
-            await asyncio.sleep(min(RETRY_BACKOFF_SEC * (2 ** (attempt - 1)), RETRY_BACKOFF_MAX_SEC))
-        try:
-            pow_headers = await _fresh_pow_headers(account)
-            resp = await _send_with_auth(
-                account,
-                account.client,
-                pow_headers,
-                session.id,
-                parent_message_id,
-                CONTINUE_PROMPT,
-                model_type,
-                thinking,
-                search,
-                ref_file_ids,
-            )
-        except HTTPException as exc:
-            if _is_retryable_http(exc) and attempt < MAX_RETRIES:
-                log.warning(
-                    "deepseek continuation error (%s), attempt %d/%d",
-                    exc.status_code,
-                    attempt + 1,
-                    MAX_RETRIES,
-                )
-                continue
-            return None
-        rec = MessageReconstructor()
-        incremental = IncrementalSSE()
-        try:
-            async for chunk in resp.aiter_bytes():
-                for event in incremental.feed(chunk):
-                    rec.handle(event)
-            for event in incremental.finish():
-                rec.handle(event)
-        finally:
-            await resp.aclose()
-        if not (rec.content or rec.reasoning) and _is_retryable_hint(rec) and attempt < MAX_RETRIES:
-            continue
-        return rec
-
-
-async def _collect_non_stream(
-    account,
-    pool,
-    existing_sid,
-    lock,
-    prompt,
-    model,
-    model_type,
-    thinking,
-    search,
-    ref_file_ids=None,
-    tool_mode=False,
-    tool_schemas=None,
-    include_usage=False,
-    context_seq: tuple[str, ...] | None = None,
-):
-    await _human_delay()
-    async with account_lock(lock, settings.acquire_timeout):
-        session, session_key, parent_message_id = await _prepare_session(account, pool, existing_sid, context_seq)
-        stop_message_id: str | None = None
-        started = time.monotonic()
-        try:
-            rec: MessageReconstructor | None = None
-            response_message_id = None
-            for attempt in range(MAX_RETRIES + 1):
-                if attempt:
-                    await asyncio.sleep(min(RETRY_BACKOFF_SEC * (2 ** (attempt - 1)), RETRY_BACKOFF_MAX_SEC))
-                try:
-                    pow_headers = await _fresh_pow_headers(account)
-                    resp = await _send_with_auth(
-                        account,
-                        account.client,
-                        pow_headers,
-                        session.id,
-                        parent_message_id,
-                        prompt,
-                        model_type,
-                        thinking,
-                        search,
-                        ref_file_ids,
-                    )
-                except HTTPException as exc:
-                    input_hint = _input_exceeds_hint_from_http(exc)
-                    if input_hint is not None:
-                        rec = MessageReconstructor()
-                        rec.hint_error = input_hint
-                        break
-                    if _is_retryable_http(exc) and attempt < MAX_RETRIES:
-                        log.warning(
-                            "deepseek provider error (%s), attempt %d/%d",
-                            exc.status_code,
-                            attempt + 1,
-                            MAX_RETRIES,
-                        )
-                        continue
-                    raise
-                rec = MessageReconstructor()
-                incremental = IncrementalSSE()
-                response_message_id = None
-                try:
-                    async for chunk in resp.aiter_bytes():
-                        for event in incremental.feed(chunk):
-                            if event.event == "ready" and isinstance(event.data, dict):
-                                response_message_id = event.data.get("response_message_id")
-                                if response_message_id:
-                                    stop_message_id = response_message_id
-                            rec.handle(event)
-                    for event in incremental.finish():
-                        rec.handle(event)
-                finally:
-                    await resp.aclose()
-                if rec.id:
-                    stop_message_id = rec.id
-                if not (rec.content or rec.reasoning) and _is_retryable_hint(rec) and attempt < MAX_RETRIES:
-                    log.warning(
-                        "retryable hint (%s), attempt %d/%d",
-                        (rec.hint_error or {}).get("finish_reason"),
-                        attempt + 1,
-                        MAX_RETRIES,
-                    )
-                    continue
-                break
-        except asyncio.CancelledError:
-            await _try_stop_stream(account.client, session.id, stop_message_id)
-            raise
-
-        assert rec is not None
-        if _is_context_limit(rec) and not (rec.content or rec.reasoning):
-            _drop_session(pool, account, session_key)
-            raise HTTPException(400, "context length exceeded: conversation too long, start a new conversation")
-        if _is_input_exceeds_limit(rec):
-            cont_parent = rec.id or response_message_id or parent_message_id
-            for _ in range(MAX_CONTINUE_ROUNDS):
-                cont_rec = await _collect_continuation(account, session, cont_parent, model_type, thinking, search, ref_file_ids)
-                if cont_rec is None:
-                    break
-                rec.extend_with(cont_rec)
-                cont_parent = cont_rec.id or cont_parent
-                if not _is_input_exceeds_limit(cont_rec):
-                    break
-        session.accumulated_tokens = max(getattr(session, "accumulated_tokens", 0) or 0, rec.accumulated_tokens)
-        account.sessions.touch_last_message(session_key, rec.id or response_message_id)
-        log.info("deepseek completion OK (%.0fms)", (time.monotonic() - started) * 1000)
-
-        if not (rec.content or rec.reasoning) and rec.hint_error:
-            raise HTTPException(429, _busy_error_body(rec))
-
-        content = rec.content
-        reasoning = rec.reasoning
-        if tool_mode:
-            parsed = toolemu.parse_tool_calls(content, tool_schemas)
-            if parsed is not None:
-                tool_calls, tool_text = parsed
-                finish = "tool_calls"
-                message = toolemu.format_tool_message(tool_calls, tool_text, reasoning)
-            else:
-                message = {"role": "assistant", "content": content}
-                if reasoning:
-                    message["reasoning_content"] = reasoning
-                finish = _finish_reason(rec.status)
-        else:
-            finish = _finish_reason(rec.status)
-            message = {"role": "assistant", "content": content}
-            if reasoning:
-                message["reasoning_content"] = reasoning
-        return {
-            "id": f"chatcmpl-{uuid.uuid4().hex}",
-            "object": "chat.completion",
-            "created": int(time.time()),
-            "model": model,
-            "choices": [
-                {
-                    "index": 0,
-                    "message": message,
-                    "finish_reason": finish,
-                }
-            ],
-            "usage": _deepseek_usage(getattr(session, "accumulated_tokens", 0)),
-            "session_id": session_key,
-        }
-
-
-async def _stream_openai(
-    account,
-    pool,
-    existing_sid,
-    lock,
-    prompt,
-    model,
-    model_type,
-    thinking,
-    search,
-    ref_file_ids=None,
-    tool_mode=False,
-    tool_schemas=None,
-    include_usage=False,
-    context_seq: tuple[str, ...] | None = None,
-):
-    chunk_id = f"chatcmpl-{uuid.uuid4().hex}"
-    created = int(time.time())
-
-    def sse(data: dict) -> str:
-        return f"data: {json.dumps(data, ensure_ascii=False)}\n\n"
-
-    await _human_delay()
-    async with account_lock(lock, settings.acquire_timeout):
-        try:
-            session, session_key, parent_message_id = await _prepare_session(account, pool, existing_sid, context_seq)
-        except HTTPException as exc:
-            detail = exc.detail if isinstance(exc.detail, str) else str(exc.detail)
-            yield sse(
-                {
-                    "id": chunk_id,
-                    "object": "chat.completion.chunk",
-                    "created": created,
-                    "model": model,
-                    "error": {"message": detail},
-                    "choices": [{"index": 0, "delta": {}, "finish_reason": "error"}],
-                }
-            )
-            yield sse(
-                {
-                    "id": chunk_id,
-                    "session_id": None,
-                    "object": "chat.completion.chunk",
-                    "choices": [],
-                }
-            )
-            yield "data: [DONE]\n\n"
-            return
-
-        rec: MessageReconstructor | None = None
-        response_message_id = None
-        stop_message_id: str | None = None
-        content_buf = ""
-        started = time.monotonic()
-        for attempt in range(MAX_RETRIES + 1):
-            if attempt:
-                await asyncio.sleep(min(RETRY_BACKOFF_SEC * (2 ** (attempt - 1)), RETRY_BACKOFF_MAX_SEC))
-            try:
-                pow_headers = await _fresh_pow_headers(account)
-                resp = await _send_with_auth(
-                    account,
-                    account.client,
-                    pow_headers,
-                    session.id,
-                    parent_message_id,
-                    prompt,
-                    model_type,
-                    thinking,
-                    search,
-                    ref_file_ids,
-                )
-            except HTTPException as exc:
-                input_hint = _input_exceeds_hint_from_http(exc)
-                if input_hint is not None:
-                    rec = MessageReconstructor()
-                    rec.hint_error = input_hint
-                    break
-                if _is_retryable_http(exc) and attempt < MAX_RETRIES:
-                    log.warning(
-                        "deepseek provider error (%s), attempt %d/%d",
-                        exc.status_code,
-                        attempt + 1,
-                        MAX_RETRIES,
-                    )
-                    continue
-                detail = exc.detail if isinstance(exc.detail, str) else str(exc.detail)
-                yield sse(
-                    {
-                        "id": chunk_id,
-                        "object": "chat.completion.chunk",
-                        "created": created,
-                        "model": model,
-                        "error": {"message": detail},
-                        "choices": [{"index": 0, "delta": {}, "finish_reason": "error"}],
-                    }
-                )
-                yield sse(
-                    {
-                        "id": chunk_id,
-                        "session_id": session_key,
-                        "object": "chat.completion.chunk",
-                        "choices": [],
-                    }
-                )
-                yield "data: [DONE]\n\n"
-                return
-            rec = MessageReconstructor()
-            incremental = IncrementalSSE()
-            response_message_id = None
-            got_content = False
-            pending: list[str] = []
-            try:
-                async for chunk in resp.aiter_bytes():
-                    for event in incremental.feed(chunk):
-                        if event.event == "ready" and isinstance(event.data, dict):
-                            response_message_id = event.data.get("response_message_id")
-                            if response_message_id:
-                                stop_message_id = response_message_id
-                        rec.handle(event)
-                        c_diff, r_diff = rec.take_diffs()
-                        if c_diff or r_diff:
-                            got_content = True
-                        if not pending:
-                            pending.append(
-                                sse(
-                                    {
-                                        "id": chunk_id,
-                                        "object": "chat.completion.chunk",
-                                        "created": created,
-                                        "model": model,
-                                        "choices": [
-                                            {
-                                                "index": 0,
-                                                "delta": {"role": "assistant"},
-                                                "finish_reason": None,
-                                            }
-                                        ],
-                                    }
-                                )
-                            )
-                        delta: dict = {}
-                        if c_diff:
-                            if tool_mode:
-                                content_buf += c_diff
-                            else:
-                                delta["content"] = c_diff
-                        if r_diff:
-                            delta["reasoning_content"] = r_diff
-                        if delta:
-                            pending.append(
-                                sse(
-                                    {
-                                        "id": chunk_id,
-                                        "object": "chat.completion.chunk",
-                                        "created": created,
-                                        "model": model,
-                                        "choices": [
-                                            {
-                                                "index": 0,
-                                                "delta": delta,
-                                                "finish_reason": None,
-                                            }
-                                        ],
-                                    }
-                                )
-                            )
-                        if got_content:
-                            for line in pending:
-                                yield line
-                            pending.clear()
-            finally:
-                if rec.id:
-                    stop_message_id = rec.id
-                if sys.exc_info()[0] is not None:
-                    await _try_stop_stream(account.client, session.id, stop_message_id)
-                await resp.aclose()
-            if got_content:
-                break
-            if _is_retryable_hint(rec) and attempt < MAX_RETRIES:
-                log.warning(
-                    "retryable hint (%s), attempt %d/%d",
-                    (rec.hint_error or {}).get("finish_reason"),
-                    attempt + 1,
-                    MAX_RETRIES,
-                )
-                continue
-            break
-
-        assert rec is not None
-        if _is_context_limit(rec) and not (rec.content or rec.reasoning):
-            _drop_session(pool, account, session_key)
-            yield sse(
-                {
-                    "id": chunk_id,
-                    "object": "chat.completion.chunk",
-                    "created": created,
-                    "model": model,
-                    "error": {
-                        "message": "context length exceeded: conversation too long, start a new conversation",
-                        "finish_reason": CONTEXT_LENGTH_STATUS,
-                    },
-                    "choices": [
-                        {
-                            "index": 0,
-                            "delta": {},
-                            "finish_reason": "length",
-                        }
-                    ],
-                }
-            )
-            yield sse(
-                {
-                    "id": chunk_id,
-                    "session_id": session_key,
-                    "object": "chat.completion.chunk",
-                    "choices": [],
-                }
-            )
-            yield "data: [DONE]\n\n"
-            return
-        if _is_input_exceeds_limit(rec):
-            cont_parent = rec.id or response_message_id or parent_message_id
-            for _ in range(MAX_CONTINUE_ROUNDS):
-                cont_rec = await _collect_continuation(account, session, cont_parent, model_type, thinking, search, ref_file_ids)
-                if cont_rec is None:
-                    break
-                rec.extend_with(cont_rec)
-                cont_parent = cont_rec.id or cont_parent
-                if cont_rec.content:
-                    if tool_mode:
-                        content_buf += cont_rec.content
-                    else:
-                        yield sse(
-                            {
-                                "id": chunk_id,
-                                "object": "chat.completion.chunk",
-                                "created": created,
-                                "model": model,
-                                "choices": [{"index": 0, "delta": {"content": cont_rec.content}, "finish_reason": None}],
-                            }
-                        )
-                if cont_rec.reasoning:
-                    yield sse(
-                        {
-                            "id": chunk_id,
-                            "object": "chat.completion.chunk",
-                            "created": created,
-                            "model": model,
-                            "choices": [{"index": 0, "delta": {"reasoning_content": cont_rec.reasoning}, "finish_reason": None}],
-                        }
-                    )
-                if not _is_input_exceeds_limit(cont_rec):
-                    break
-        if not (rec.content or rec.reasoning) and rec.hint_error:
-            hint = rec.hint_error
-            yield sse(
-                {
-                    "id": chunk_id,
-                    "object": "chat.completion.chunk",
-                    "created": created,
-                    "model": model,
-                    "error": {
-                        "message": hint.get("message") or "DeepSeek server is busy, try again later",
-                        "finish_reason": hint.get("finish_reason"),
-                    },
-                    "choices": [
-                        {
-                            "index": 0,
-                            "delta": {},
-                            "finish_reason": hint.get("finish_reason") or "error",
-                        }
-                    ],
-                }
-            )
-            yield sse(
-                {
-                    "id": chunk_id,
-                    "session_id": session_key,
-                    "object": "chat.completion.chunk",
-                    "choices": [],
-                }
-            )
-            yield "data: [DONE]\n\n"
-            return
-
-        session.accumulated_tokens = max(getattr(session, "accumulated_tokens", 0) or 0, rec.accumulated_tokens)
-        account.sessions.touch_last_message(session_key, rec.id or response_message_id)
-        log.info("deepseek completion OK (%.0fms)", (time.monotonic() - started) * 1000)
-
-        if tool_mode:
-            parsed = toolemu.parse_tool_calls(content_buf, tool_schemas)
-            if parsed is not None:
-                tool_calls, tool_text = parsed
-                for delta in toolemu.tool_call_deltas(tool_calls, tool_text):
-                    yield sse(
-                        {
-                            "id": chunk_id,
-                            "object": "chat.completion.chunk",
-                            "created": created,
-                            "model": model,
-                            "choices": [{"index": 0, "delta": delta, "finish_reason": None}],
-                        }
-                    )
-                finish = "tool_calls"
-            else:
-                if content_buf:
-                    yield sse(
-                        {
-                            "id": chunk_id,
-                            "object": "chat.completion.chunk",
-                            "created": created,
-                            "model": model,
-                            "choices": [
-                                {
-                                    "index": 0,
-                                    "delta": {"content": content_buf},
-                                    "finish_reason": None,
-                                }
-                            ],
-                        }
-                    )
-                finish = _finish_reason(rec.status)
-        else:
-            finish = _finish_reason(rec.status)
-
-        yield sse(
-            {
-                "id": chunk_id,
-                "object": "chat.completion.chunk",
-                "created": created,
-                "model": model,
-                "choices": [{"index": 0, "delta": {}, "finish_reason": finish}],
-            }
-        )
-        if include_usage:
-            yield sse(
-                {
-                    "id": chunk_id,
-                    "object": "chat.completion.chunk",
-                    "created": created,
-                    "model": model,
-                    "choices": [],
-                    "usage": _deepseek_usage(getattr(session, "accumulated_tokens", 0)),
-                }
-            )
-        yield sse(
-            {
-                "id": chunk_id,
-                "session_id": session_key,
-                "object": "chat.completion.chunk",
-                "choices": [],
-            }
-        )
-        yield "data: [DONE]\n\n"
+from .attachments import (  # noqa: E402,F401
+    MAX_FILE_SIZE,
+    MAX_FILES_PER_REQUEST,
+    Attachment,
+    _collect_attachments,
+    _fresh_pow_upload_headers,
+    _split_data_uri,
+    _upload_attachments,
+    _validate_attachments,
+)
+from .deepseek import (  # noqa: E402,F401
+    DEEPSEEK_AUTH_ERROR_CODES,
+    MAX_RETRIES,
+    RETRY_BACKOFF_MAX_SEC,
+    RETRY_BACKOFF_SEC,
+    RETRYABLE_FINISH_REASONS,
+    _busy_error_body,
+    _collect_continuation,
+    _deepseek_error_detail,
+    _deepseek_status,
+    _drop_session,
+    _fresh_pow_headers,
+    _handle_account_error,
+    _human_delay,
+    _input_exceeds_hint_from_http,
+    _is_context_limit,
+    _is_input_exceeds_limit,
+    _is_retryable_hint,
+    _is_retryable_http,
+    _prepare_session,
+    _send_completion,
+    _send_with_auth,
+    _try_stop_stream,
+)
+from .models import (  # noqa: E402,F401
+    MODEL_TYPE_BY_NAME,
+    QWEN_DEFAULT_MODELS,
+    ChatMessage,
+    FileSpec,
+    _deepseek_usage,
+    _finish_reason,
+    _include_usage,
+    _resolve_model,
+)
+from .streaming import (  # noqa: E402,F401
+    _collect_non_stream,
+    _sse,
+    _stream_guard,
+    _stream_openai,
+)
